refactor bgen and remove unstaged index writer
