+++ hail/src/main/scala/is/hail/io/index/InternalNodeBuilder.scala
@@ -31,61 +31,3 @@ object InternalNodeBuilder {
     "children" -> arrayType(keyType, annotationType)
   )
 }
-
-class InternalNodeBuilder(keyType: PType, annotationType: PType) {
-  val indexFileOffsets = new ArrayBuilder[Long]()
-  val firstIdxs = new ArrayBuilder[Long]()
-  val firstKeys = new ArrayBuilder[Any]()
-  val firstRecordOffsets = new ArrayBuilder[Long]()
-  val firstAnnotations = new ArrayBuilder[Any]()
-
-  var size = 0
-  val pType: PStruct = InternalNodeBuilder.typ(keyType, annotationType)
-
-  def +=(info: IndexNodeInfo) {
-    indexFileOffsets += info.indexFileOffset
-    firstIdxs += info.firstIndex
-    firstKeys += info.firstKey
-    firstRecordOffsets += info.firstRecordOffset
-    firstAnnotations += info.firstAnnotation
-    size += 1
-  }
-
-  def write(rvb: RegionValueBuilder): Long = {
-    rvb.start(pType)
-    rvb.startStruct()
-    rvb.startArray(size)
-    var i = 0
-    while (i < size) {
-      rvb.startStruct()
-      rvb.addLong(indexFileOffsets(i))
-      rvb.addLong(firstIdxs(i))
-      rvb.addAnnotation(keyType.virtualType, firstKeys(i))
-      rvb.addLong(firstRecordOffsets(i))
-      rvb.addAnnotation(annotationType.virtualType, firstAnnotations(i))
-      rvb.endStruct()
-      i += 1
-    }
-    rvb.endArray()
-    rvb.endStruct()
-    rvb.end()
-  }
-
-  def clear() {
-    indexFileOffsets.clear()
-    firstIdxs.clear()
-    firstKeys.clear()
-    firstRecordOffsets.clear()
-    firstAnnotations.clear()
-    size = 0
-  }
-
-  def getChild(idx: Int): InternalChild = {
-    assert(idx >= 0 && idx < size)
-    InternalChild(indexFileOffsets(idx), firstIdxs(idx), firstKeys(idx), firstRecordOffsets(idx), firstAnnotations(idx))
-  }
-
-  override def toString: String = s"InternalNodeBuilder $size [${ (0 until size).map { i =>
-    (indexFileOffsets(i), firstIdxs(i), firstKeys(i), firstRecordOffsets(i), firstAnnotations(i))
-  }.mkString(",") }]"
-}
